feat(core, llm): add supporting tags
Controllers now can pass two set of tags to the downstream Pulumi
resources. First is the user defined tags which is a list of tags set at
initialization of the controller. Second is a set of default tags which
is enforced by the controller. By default damavand's base controller is
setting two default tags including `application` and `tool`.